requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
yfinance>=0.2.3
pandas>=2.0.0
numpy>=1.24.0
python-ta>=0.10.0  # For technical analysis 
TA-Lib>=0.4.24  # Technical analysis library 